Simplify commands in README
I think the README should be as simple as possible and focus on useful information. It is obvious that this is a Bash script, which should thus be called with `./mtc` or `bash mtc`. Moreover, using the `-h` flag is standard practice, and as such does not need to be repeated for each command.